move entities from domain to businesslogic
