--- server/Application/UserManagement/DTOs/RegisterUserDTO.cs
@@ -1,10 +1,0 @@
-namespace CoffeeShop.Application.UserManagement.DTOs;
-
-public sealed record RegisterUserDTO(
-    string FirstName,
-    string LastName,
-    string Email,
-    string PhoneNumber,
-    string Password,
-    string Role
-);
