--- server/Application/UserManagement/Interfaces/IUserService.cs
@@ -1,8 +1,0 @@
-using CoffeeShop.Application.UserManagement.DTOs;
-
-namespace CoffeeShop.Application.UserManagement.Interfaces;
-
-public interface IUserService
-{
-    Task<UserDTO> CreateUser(RegisterUserDTO dto);
-}
