+++ server/BusinessLogic/UserManagement/Enums/Roles.cs
@@ -1,0 +1,8 @@
+namespace CoffeeShop.BusinessLogic.UserManagement.Enums;
+
+public enum Roles
+{
+    BusinessOwner = 0,
+    Employee,
+    Client,
+}
