+++ server/BusinessLogic/UserManagement/Interfaces/IUserService.cs
@@ -1,0 +1,8 @@
+using CoffeeShop.BusinessLogic.UserManagement.Entities;
+
+namespace CoffeeShop.BusinessLogic.UserManagement.Interfaces;
+
+public interface IUserService
+{
+    Task<User> AddUser(User user, string password, string role);
+}
