+++ server/BusinessLogic/UserManagement/Services/UserService.cs
@@ -1,0 +1,33 @@
+using System.ComponentModel.DataAnnotations;
+using CoffeeShop.BusinessLogic.UserManagement.Entities;
+using CoffeeShop.BusinessLogic.UserManagement.Interfaces;
+using Microsoft.AspNetCore.Identity;
+
+namespace CoffeeShop.BusinessLogic.UserManagement.Services;
+
+public class UserService : IUserService
+{
+    private readonly UserManager<User> _userManager;
+
+    public UserService(UserManager<User> userManager)
+    {
+        _userManager = userManager;
+    }
+
+    public async Task<User> AddUser(User user, string password, string role)
+    {
+        var result = await _userManager.CreateAsync(user, password);
+
+        if (result.Succeeded)
+        {
+            await _userManager.AddToRoleAsync(user, role);
+        }
+        else
+        {
+            var errors = string.Join(", ", result.Errors.Select(e => e.Description));
+            throw new ValidationException(errors);
+        }
+
+        return user;
+    }
+}
